requirements.txt
django
djangorestframework
djangorestframework-simplejwt    
django-cors-headers              
django-filter                   
psycopg2-binary                 
python-decouple  
django-cors-headers